fix: command timeout cancel
